hw-06-jpql/src/main/java/ru/otus/hw06jpql/repository/AuthorRepositoryJpa.java
package ru.otus.hw06jpql.repository;

import org.springframework.stereotype.Repository;
import ru.otus.hw06jpql.domain.Author;

import javax.persistence.EntityManager;
import javax.persistence.PersistenceContext;
import java.util.Optional;

@Repository
public class AuthorRepositoryJpa implements AuthorRepository {

    @PersistenceContext
    private EntityManager em;

    @Override
    public Author insertOrUpdate(Author author) {
        if (author.getId() <= 0) {
            em.persist(author);
            return author;
        } else {
            return em.merge(author);
        }
    }

    @Override
    public Optional<Author> getById(long id) {
        return Optional.ofNullable(em.find(Author.class, id));
    }
}
